+++ click_extended/transform/__init__.py
@@ -1,7 +1,11 @@
 """Initialization file for the `click_extended.transform` module."""
 
 from click_extended.transform.as_path import as_path
+from click_extended.transform.to_lowercase import to_lowercase
+from click_extended.transform.to_uppercase import to_uppercase
 
 __all__ = [
     "as_path",
+    "to_lowercase",
+    "to_uppercase",
 ]